separate main menu from Makelangelo class
