+++ src/main/java/com/marginallyclever/makelangelo/Makelangelo.java
@@ -2,40 +2,22 @@ package com.marginallyclever.makelangelo;
 
 import com.formdev.flatlaf.FlatLaf;
 import com.formdev.flatlaf.FlatLightLaf;
-import com.hopding.jrpicam.exceptions.FailedToRunRaspistillException;
 import com.marginallyclever.convenience.CommandLineOptions;
 import com.marginallyclever.convenience.FileAccess;
-import com.marginallyclever.convenience.StringHelper;
 import com.marginallyclever.convenience.log.Log;
-import com.marginallyclever.convenience.log.LogPanel;
-import com.marginallyclever.makelangelo.firmwareuploader.FirmwareUploaderPanel;
-import com.marginallyclever.makelangelo.makeart.TurtleModifierAction;
 import com.marginallyclever.makelangelo.makeart.io.LoadFilePanel;
-import com.marginallyclever.makelangelo.makeart.io.OpenFileChooser;
 import com.marginallyclever.makelangelo.makeart.io.SaveGCode;
-import com.marginallyclever.makelangelo.makeart.tools.*;
-import com.marginallyclever.makelangelo.makeart.turtlegenerator.TurtleGenerator;
-import com.marginallyclever.makelangelo.makeart.turtlegenerator.TurtleGeneratorFactory;
-import com.marginallyclever.makelangelo.makeart.turtlegenerator.TurtleGeneratorPanel;
-import com.marginallyclever.makelangelo.makelangelosettingspanel.GFXPreferences;
 import com.marginallyclever.makelangelo.makelangelosettingspanel.LanguagePreferences;
-import com.marginallyclever.makelangelo.makelangelosettingspanel.MakelangeloSettingPanel;
 import com.marginallyclever.makelangelo.makelangelosettingspanel.MetricsPreferences;
 import com.marginallyclever.makelangelo.paper.Paper;
-import com.marginallyclever.makelangelo.paper.PaperSettings;
-import com.marginallyclever.makelangelo.plotter.PiCaptureAction;
 import com.marginallyclever.makelangelo.plotter.Plotter;
-import com.marginallyclever.makelangelo.plotter.marlinsimulation.MarlinSimulation;
 import com.marginallyclever.makelangelo.plotter.marlinsimulation.MarlinSimulationVisualizer;
-import com.marginallyclever.makelangelo.plotter.plottercontrols.PlotterControls;
 import com.marginallyclever.makelangelo.plotter.plotterrenderer.Machines;
 import com.marginallyclever.makelangelo.plotter.plotterrenderer.PlotterRenderer;
 import com.marginallyclever.makelangelo.plotter.plottersettings.PlotterSettings;
 import com.marginallyclever.makelangelo.plotter.plottersettings.PlotterSettingsManager;
-import com.marginallyclever.makelangelo.plotter.plottersettings.PlotterSettingsManagerPanel;
 import com.marginallyclever.makelangelo.preview.Camera;
 import com.marginallyclever.makelangelo.preview.PreviewPanel;
-import com.marginallyclever.makelangelo.rangeslider.RangeSlider;
 import com.marginallyclever.makelangelo.turtle.Turtle;
 import com.marginallyclever.makelangelo.turtle.turtlerenderer.TurtleRenderFacade;
 import com.marginallyclever.makelangelo.turtle.turtlerenderer.TurtleRenderFactory;
@@ -46,11 +28,8 @@ import org.slf4j.LoggerFactory;
 
 import javax.imageio.ImageIO;
 import javax.swing.*;
-import javax.swing.event.ChangeEvent;
 import java.awt.*;
 import java.awt.dnd.DropTarget;
-import java.awt.event.InputEvent;
-import java.awt.event.KeyEvent;
 import java.awt.event.WindowAdapter;
 import java.awt.event.WindowEvent;
 import java.io.BufferedReader;
@@ -58,10 +37,8 @@ import java.io.File;
 import java.io.IOException;
 import java.io.InputStreamReader;
 import java.net.HttpURLConnection;
-import java.net.URI;
 import java.net.URL;
 import java.security.InvalidParameterException;
-import java.util.Arrays;
 import java.util.Locale;
 import java.util.Objects;
 import java.util.prefs.Preferences;
@@ -81,40 +58,25 @@ import java.util.prefs.Preferences;
  */
 public final class Makelangelo {
 	private static final String PREFERENCE_SAVE_PATH = "savePath";
-	private static int SHORTCUT_CTRL = InputEvent.CTRL_DOWN_MASK;
-	private static int SHORTCUT_ALT = InputEvent.ALT_DOWN_MASK;
-
 	private static Logger logger;
-
-	private final MakelangeloSettingPanel myPreferencesPanel = new MakelangeloSettingPanel();
-	
 	private final Camera camera;
-
 	private final PlotterSettingsManager plotterSettingsManager = new PlotterSettingsManager();
 	private final Plotter myPlotter = new Plotter();
 	private final Paper myPaper = new Paper();
 	private Turtle myTurtle = new Turtle();
-	private static boolean isMacOS = false;
-
 	private final TurtleRenderFacade myTurtleRenderer = new TurtleRenderFacade();
-	
 	private PlotterRenderer myPlotterRenderer;
 	
 	// GUI elements
 	private MainFrame mainFrame;
-	private JMenuBar mainMenuBar;
+	private MainMenu mainMenuBar = new MainMenu(this);
 	private PreviewPanel previewPanel;
-	private final SaveDialog saveDialog = new SaveDialog();
-	
-	private RecentFiles recentFiles;
-	private OpenFileChooser openFileChooser;
-
-	private RangeSlider rangeSlider;
-	private final JLabel labelRangeMin = new JLabel();
-	private final JLabel labelRangeMax = new JLabel();
+	private final MakeleangeloRangeSlider rangeSlider = new MakeleangeloRangeSlider();
 
 
 	public Makelangelo() {
+		super();
+
 		logger.debug("Locale={}", Locale.getDefault().toString());
 		logger.debug("Headless={}", (GraphicsEnvironment.isHeadless()?"Y":"N"));
 
@@ -134,6 +96,11 @@ public final class Makelangelo {
 			addPlotterRendererToPreviewPanel();
 		}
 
+		rangeSlider.addChangeListener(e->{
+			myTurtleRenderer.setFirst(rangeSlider.getBottom());
+			myTurtleRenderer.setLast(rangeSlider.getTop());
+		});
+
 		onPlotterSettingsUpdate(myPlotter.getSettings());
 	}
 
@@ -146,7 +113,7 @@ public final class Makelangelo {
 		}
 	}
 
-	private void onPlotterSettingsUpdate(PlotterSettings e) {
+	public void onPlotterSettingsUpdate(PlotterSettings e) {
 		myPlotter.setSettings(e);
 		if(previewPanel != null) previewPanel.repaint();
 		TurtleRenderer f = TurtleRenderFactory.MARLIN_SIM.getTurtleRenderer();
@@ -178,7 +145,6 @@ public final class Makelangelo {
 
 	private static void setSystemLookAndFeel() {
 		if(!CommandLineOptions.hasOption("-nolf")) {
-
 			try {
 				FlatLaf.registerCustomDefaultsSource( "com.marginallyclever.makelangelo" );
 				UIManager.setLookAndFeel( new FlatLightLaf() );
@@ -190,17 +156,6 @@ public final class Makelangelo {
 					logger.warn("failed to set native look and feel.", ex);
 				}
 			}
-		}
-		setSystemLookAndFeelForMacos();
-	}
-	
-	private static void setSystemLookAndFeelForMacos() {
-		String os = System.getProperty("os.name").toLowerCase(Locale.ENGLISH);
-		if ((os.contains("mac")) || (os.contains("darwin"))) {
-			isMacOS = true;
-			System.setProperty("apple.laf.useScreenMenuBar", "true");
-			SHORTCUT_CTRL = InputEvent.META_DOWN_MASK;
-			SHORTCUT_ALT = InputEvent.META_DOWN_MASK;
 		}
 	}
 
@@ -222,355 +177,15 @@ public final class Makelangelo {
 	}
 
 	/**
-	 * Build the main menu
-	 */
-	private void buildMenuBar() {
-		logger.debug("  adding menu bar...");
-
-		mainMenuBar = new JMenuBar();
-		mainMenuBar.add(createFileMenu());
-		mainMenuBar.add(createSettingsMenu());
-		mainMenuBar.add(createGenerateMenu());
-		mainMenuBar.add(createToolsMenu());
-		mainMenuBar.add(createViewMenu());
-		mainMenuBar.add(createRobotMenu());
-		mainMenuBar.add(createHelpMenu());
-		mainMenuBar.updateUI();
-		
-		mainFrame.setJMenuBar(mainMenuBar);
-	}
-
-	/**
 	 * Change the enable state of the menu items inside the {@code mainMenuBar}.
 	 * Remember that enabling the menuBar does not affect the children.
 	 * @param b the new state
 	 */
-	private void enableMenuBar(boolean b) {
+	public void enableMenuBar(boolean b) {
 		int c = mainMenuBar.getMenuCount();
 		while(--c>=0) {
 			mainMenuBar.getMenu(c).setEnabled(b);
 		}
-	}
-
-	private JMenu createSettingsMenu() {
-		JMenu menu = new JMenu(Translator.get("MenuSettings"));
-		menu.setMnemonic('S');
-		
-		JMenuItem bOpenPaperSettings = new JMenuItem(Translator.get("OpenPaperSettings"));
-		bOpenPaperSettings.addActionListener((e)-> openPaperSettings());
-		menu.add(bOpenPaperSettings);
-
-		JMenuItem bOpenPlotterSettings = new JMenuItem(Translator.get("OpenPlotterSettings"));
-		bOpenPlotterSettings.addActionListener((e)-> openPlotterSettings());
-		bOpenPlotterSettings.setAccelerator(KeyStroke.getKeyStroke(KeyEvent.VK_P, SHORTCUT_CTRL));//"ctrl P"
-		menu.add(bOpenPlotterSettings);
-
-		return menu;
-	}
-
-	private void openPlotterSettings() {
-		PlotterSettingsManagerPanel plotterSettingsPanel = new PlotterSettingsManagerPanel(plotterSettingsManager);
-		JDialog dialog = new JDialog(mainFrame,Translator.get("PlotterSettingsPanel.Title"));
-		dialog.add(plotterSettingsPanel);
-		dialog.setMinimumSize(new Dimension(350,300));
-		dialog.setResizable(false);
-		dialog.pack();
-
-		enableMenuBar(false);
-		dialog.addWindowListener(new WindowAdapter() {
-			@Override
-			public void windowClosing(WindowEvent e) {
-				onPlotterSettingsUpdate(plotterSettingsManager.getLastSelectedProfile());
-				enableMenuBar(true);
-			}
-		});
-
-		dialog.setLocationRelativeTo(mainFrame);
-		dialog.setVisible(true);
-	}
-
-	private void openPaperSettings() {
-		PaperSettings settings = new PaperSettings(myPaper);
-		JDialog dialog = new JDialog(mainFrame,Translator.get("PaperSettings.Title"));
-		dialog.add(settings);
-		dialog.setMinimumSize(new Dimension(300,300));
-		dialog.pack();
-
-		enableMenuBar(false);
-		dialog.addWindowListener(new WindowAdapter() {
-			@Override
-			public void windowClosing(WindowEvent e) {
-				enableMenuBar(true);
-				settings.save();
-			}
-		});
-
-		dialog.setLocationRelativeTo(mainFrame);
-		dialog.setVisible(true);
-	}
-
-	private JMenu createRobotMenu() {
-		JMenu menu = new JMenu(Translator.get("Robot"));
-		menu.setMnemonic('k');
-		
-		JMenuItem bEstimate = new JMenuItem(Translator.get("RobotMenu.GetTimeEstimate"));
-		bEstimate.setAccelerator(KeyStroke.getKeyStroke(KeyEvent.VK_E, SHORTCUT_CTRL));//"ctrl E"
-		bEstimate.addActionListener((e)-> estimateTime());
-		menu.add(bEstimate);
-
-		JMenuItem bSaveToSD = new JMenuItem(Translator.get("RobotMenu.SaveGCode"));
-		bSaveToSD.addActionListener((e)-> saveGCode());
-		bSaveToSD.setAccelerator(KeyStroke.getKeyStroke(KeyEvent.VK_G, SHORTCUT_CTRL));//"ctrl G"
-		menu.add(bSaveToSD);
-
-		JMenuItem bOpenControls = new JMenuItem(Translator.get("RobotMenu.OpenControls"));
-		bOpenControls.addActionListener((e)-> openPlotterControls());
-		bOpenControls.setAccelerator(KeyStroke.getKeyStroke(KeyEvent.VK_T, SHORTCUT_CTRL));//"ctrl C"
-		bOpenControls.setIcon(new UnicodeIcon("\uD83D\uDD79"));
-		menu.add(bOpenControls);
-
-		return menu;
-	}
-
-	private JMenuItem createRenderStyleMenu() {
-		JMenu menu = new JMenu(Translator.get("RobotMenu.RenderStyle"));
-
-		ButtonGroup group = new ButtonGroup();
-
-		Arrays.stream(TurtleRenderFactory.values())
-				.forEach(iter -> {
-					TurtleRenderer renderer = iter.getTurtleRenderer();
-					String name = iter.getName();
-					JRadioButtonMenuItem button = new JRadioButtonMenuItem(iter.getTranslatedText());
-					if (myTurtleRenderer.getRenderer() == renderer) button.setSelected(true);
-					button.addActionListener((e)-> onTurtleRenderChange(name));
-					menu.add(button);
-					group.add(button);
-				});
-
-		return menu;
-	}
-
-	private void onTurtleRenderChange(String name) {
-		logger.debug("Switching to render style '{}'", name);
-		TurtleRenderer renderer = TurtleRenderFactory.findByName(name).getTurtleRenderer();
-		myTurtleRenderer.setRenderer(renderer);
-	}
-
-	private void saveGCode() {
-		logger.debug("Saving to gcode...");
-
-		SaveGCode save = new SaveGCode();
-		try {
-			int head = rangeSlider.getValue();
-			int tail = rangeSlider.getUpperValue();
-			save.run(myTurtle, myPlotter, mainFrame, head, tail);
-		} catch(Exception e) {
-			logger.error("Error while exporting the gcode", e);
-			JOptionPane.showMessageDialog(mainFrame, Translator.get("SaveError") + e.getLocalizedMessage(), Translator.get("ErrorTitle"), JOptionPane.ERROR_MESSAGE);
-		}
-	}
-
-	private void estimateTime() {
-		MarlinSimulation ms = new MarlinSimulation(myPlotter.getSettings());
-		int estimatedSeconds = (int)Math.ceil(ms.getTimeEstimate(myTurtle));
-		String timeAsString = StringHelper.getElapsedTime(estimatedSeconds);
-		String message = Translator.get("EstimatedTimeIs",new String[]{timeAsString});
-		JOptionPane.showMessageDialog(mainFrame, message, Translator.get("RobotMenu.GetTimeEstimate"), JOptionPane.INFORMATION_MESSAGE);
-	}
-
-	private void openPlotterControls() {
-		JDialog dialog = new JDialog(mainFrame, Translator.get("PlotterControls.Title"));
-		dialog.setPreferredSize(new Dimension(PlotterControls.DIMENSION_PANEL_WIDTH, PlotterControls.DIMENSION_PANEL_HEIGHT));
-		dialog.setMinimumSize(new Dimension(PlotterControls.DIMENSION_PANEL_WIDTH, PlotterControls.DIMENSION_PANEL_HEIGHT));
-		PlotterControls plotterControls = new PlotterControls(myPlotter,myTurtle, dialog);
-		dialog.add(plotterControls);
-		dialog.pack();
-
-		enableMenuBar(false);
-		dialog.addWindowListener(new WindowAdapter() {
-			@Override
-			public void windowClosing(WindowEvent e) {
-				plotterControls.onDialogClosing();
-				enableMenuBar(true);
-			}
-		});
-
-		dialog.setLocationRelativeTo(mainFrame);
-		dialog.setVisible(true);
-	}
-
-	private JMenu createToolsMenu() {
-		JMenu menu = new JMenu(Translator.get("Art Pipeline"));
-		menu.setMnemonic('T');
-
-		try {
-			PiCaptureAction pc = new PiCaptureAction();
-
-			JButton bCapture = new JButton(Translator.get("MenuCaptureImage"));
-			bCapture.addActionListener((e)-> pc.run(mainFrame,myPaper));
-			bCapture.setIcon(new UnicodeIcon("📷"));
-			menu.add(bCapture);
-			menu.addSeparator();
-		} catch (FailedToRunRaspistillException e) {
-			logger.debug("PiCaptureAction unavailable.");
-		}
-
-		TurtleModifierAction a6 = new ResizeTurtleToPaperAction(myPaper,false,Translator.get("ConvertImagePaperFit"));
-		TurtleModifierAction a7 = new ResizeTurtleToPaperAction(myPaper,true,Translator.get("ConvertImagePaperFill"));
-		a6.setSource(this);		a6.addModifierListener(this::setTurtle);		menu.add(a6);
-		a7.setSource(this);		a7.addModifierListener(this::setTurtle);		menu.add(a7);
-		
-		JMenuItem scale = new JMenuItem(Translator.get("Scale"));
-		menu.add(scale);
-		scale.addActionListener((e) -> runScalePanel());
-
-		JMenuItem rotate = new JMenuItem(Translator.get("Rotate"));
-		rotate.setIcon(new UnicodeIcon("↻"));
-		menu.add(rotate);
-		rotate.addActionListener((e) -> runRotatePanel());
-		menu.addSeparator();
-		
-		TurtleModifierAction a4 = new FlipTurtleAction(1,-1,Translator.get("FlipH"));
-		a4.putValue(Action.SMALL_ICON, new UnicodeIcon("↕"));
-		a4.putValue(Action.ACCELERATOR_KEY, KeyStroke.getKeyStroke(java.awt.event.KeyEvent.VK_H, SHORTCUT_CTRL));//"ctrl H"
-		TurtleModifierAction a5 = new FlipTurtleAction(-1,1,Translator.get("FlipV"));
-		a5.putValue(Action.SMALL_ICON, new UnicodeIcon("↔"));
-		a5.putValue(Action.ACCELERATOR_KEY, KeyStroke.getKeyStroke(java.awt.event.KeyEvent.VK_F, SHORTCUT_CTRL));//"ctrl F"
-		a4.setSource(this);		a4.addModifierListener(this::setTurtle);		menu.add(a4);
-		a5.setSource(this);		a5.addModifierListener(this::setTurtle);		menu.add(a5);
-		
-		menu.addSeparator();
-		
-		TurtleModifierAction a1 = new SimplifyTurtleAction();
-		a1.putValue(Action.ACCELERATOR_KEY, KeyStroke.getKeyStroke(java.awt.event.KeyEvent.VK_Y, SHORTCUT_CTRL));//"ctrl Y"
-		TurtleModifierAction a2 = new ReorderTurtleAction();
-		a2.putValue(Action.ACCELERATOR_KEY, KeyStroke.getKeyStroke(java.awt.event.KeyEvent.VK_R, SHORTCUT_CTRL));//"ctrl R"
-		TurtleModifierAction a3 = new InfillTurtleAction();
-		a3.putValue(Action.ACCELERATOR_KEY, KeyStroke.getKeyStroke(java.awt.event.KeyEvent.VK_I, SHORTCUT_CTRL));//"ctrl I"
-		a1.setSource(this);		a1.addModifierListener(this::setTurtle);		menu.add(a1);
-		a2.setSource(this);		a2.addModifierListener(this::setTurtle);		menu.add(a2);
-		a3.setSource(this);		a3.addModifierListener(this::setTurtle);		menu.add(a3);
-
-		return menu;
-	}
-
-	private void runRotatePanel() {
-		RotateTurtlePanel.runAsDialog(mainFrame, myTurtle);
-	}
-	private void runScalePanel() {
-		ScaleTurtlePanel.runAsDialog(mainFrame, myTurtle);
-	}
-
-	private JMenu createGenerateMenu() {
-		JMenu menu = new JMenu(Translator.get("MenuGenerate"));
-		menu.setMnemonic('A');
-		for( TurtleGenerator ici : TurtleGeneratorFactory.available ) {
-			JMenuItem mi = new JMenuItem(ici.getName());
-			mi.addActionListener((e) -> runGeneratorDialog(ici));
-			menu.add(mi);
-		}
-		
-		return menu;
-	}
-	
-	private void runGeneratorDialog(TurtleGenerator turtleGenerator) {
-		turtleGenerator.setPaper(myPaper);
-		turtleGenerator.addListener(this::setTurtle);
-		turtleGenerator.generate();
-
-		if(turtleGenerator.getPanelElements().isEmpty()) {
-			return;
-		}
-
-		setMainTitle(turtleGenerator.getName());
-		JDialog dialog = new JDialog(mainFrame, turtleGenerator.getName());
-		TurtleGeneratorPanel panel = new TurtleGeneratorPanel(turtleGenerator);
-		dialog.add(panel);
-		dialog.setLocationRelativeTo(mainFrame);
-		dialog.setMinimumSize(new Dimension(300,300));
-		dialog.pack();
-
-
-		enableMenuBar(false);
-		dialog.addWindowListener(new WindowAdapter() {
-			@Override
-			public void windowClosing(WindowEvent e) {
-				enableMenuBar(true);
-				myPaper.setRotationRef(0);
-				logger.debug("Generation finished");
-			}
-		});
-
-		dialog.setVisible(true);
-	}
-	
-	private void newFile() {
-		setTurtle(new Turtle());
-		setMainTitle("");
-	}
-
-	private JMenu createFileMenu() {
-		JMenu menu = new JMenu(Translator.get("MenuMakelangelo"));
-		menu.setMnemonic('f');
-
-		JMenuItem buttonNewFile = new JMenuItem(Translator.get("MenuNewFile"));
-		buttonNewFile.addActionListener((e) -> newFile());
-		buttonNewFile.setAccelerator(KeyStroke.getKeyStroke(KeyEvent.VK_N, SHORTCUT_CTRL));//"ctrl N"
-		buttonNewFile.setIcon(new UnicodeIcon("\uD83C\uDF31"));
-		menu.add(buttonNewFile);
-
-		JMenuItem buttonOpenFile = new JMenuItem(Translator.get("MenuOpenFile"));
-		buttonOpenFile.addActionListener((e) -> openLoadFile());
-		buttonOpenFile.setAccelerator(KeyStroke.getKeyStroke(KeyEvent.VK_O, SHORTCUT_CTRL));//"ctrl O"
-		buttonOpenFile.setIcon(new UnicodeIcon("\uD83D\uDDC1"));
-		menu.add(buttonOpenFile);
-		
-		recentFiles = new RecentFiles(Translator.get("MenuReopenFile"));
-		recentFiles.addSubmenuListener((e)-> openFile(((JMenuItem)e.getSource()).getText()));
-		menu.add(recentFiles);
-		
-		JMenuItem buttonSaveFile = new JMenuItem(Translator.get("MenuSaveFile"));
-		buttonSaveFile.addActionListener((e) -> saveFile());
-		buttonSaveFile.setAccelerator(KeyStroke.getKeyStroke(KeyEvent.VK_S, SHORTCUT_CTRL));//"ctrl S"
-		buttonSaveFile.setIcon(new UnicodeIcon("\uD83D\uDCBE"));
-		menu.add(buttonSaveFile);
-
-		menu.addSeparator();
-				
-		JMenuItem buttonAdjustPreferences = new JMenuItem(Translator.get("MenuPreferences"));
-		buttonAdjustPreferences.addActionListener((e)-> myPreferencesPanel.run(mainFrame));
-		if (isMacOS) {
-			buttonAdjustPreferences.setAccelerator(KeyStroke.getKeyStroke(KeyEvent.VK_COMMA, SHORTCUT_CTRL));//"cmd ,"
-		} else {
-			buttonAdjustPreferences.setAccelerator(KeyStroke.getKeyStroke(KeyEvent.VK_P, SHORTCUT_ALT));//"alt P"
-		}
-		buttonAdjustPreferences.setIcon(new UnicodeIcon("⚙"));
-		menu.add(buttonAdjustPreferences);
-
-		JMenuItem buttonFirmwareUpdate = new JMenuItem(Translator.get("FirmwareUpdate"));
-		buttonFirmwareUpdate.addActionListener((e) -> runFirmwareUpdate());
-		buttonFirmwareUpdate.setIcon(new UnicodeIcon("⬆"));
-		menu.add(buttonFirmwareUpdate);
-
-		if (!isMacOS) {
-			menu.addSeparator();
-
-			JMenuItem buttonExit = new JMenuItem(Translator.get("MenuQuit"));
-			buttonExit.addActionListener((e) -> {
-				WindowEvent windowClosing = new WindowEvent(mainFrame, WindowEvent.WINDOW_CLOSING);
-				Toolkit.getDefaultToolkit().getSystemEventQueue().postEvent(windowClosing);
-			});
-			buttonExit.setAccelerator(KeyStroke.getKeyStroke(KeyEvent.VK_Q, SHORTCUT_CTRL));//"ctrl Q"
-			buttonExit.setMnemonic('Q');
-			menu.add(buttonExit);
-		}
-		return menu;
-	}
-
-	public void openLoadFile() {
-		logger.debug("Open file...");
-		openFileChooser.chooseFile();
 	}
 
 	public void openFile(String filename) {
@@ -596,13 +211,13 @@ public final class Makelangelo {
 						loader.loadingFinished();
 						enableMenuBar(true);
 						previewPanel.removeListener(loader);
-						recentFiles.addFilename(filename);
+						mainMenuBar.getRecentFiles().addFilename(filename);
 					}
 				});
 
 				dialog.setVisible(true);
 			} else {
-				recentFiles.addFilename(filename);
+				mainMenuBar.getRecentFiles().addFilename(filename);
 			}
 
 			setMainTitle(new File(filename).getName());
@@ -610,132 +225,20 @@ public final class Makelangelo {
 		} catch(Exception e) {
 			logger.error("Error while loading the file {}", filename, e);
 			JOptionPane.showMessageDialog(mainFrame, Translator.get("LoadError") + e.getLocalizedMessage(), Translator.get("ErrorTitle"), JOptionPane.ERROR_MESSAGE);
-			recentFiles.removeFilename(filename);
+			mainMenuBar.getRecentFiles().removeFilename(filename);
 		}
 	}
 
-	private void runFirmwareUpdate() {
-		JDialog dialog = new JDialog(mainFrame,"Firmware Update");
-		dialog.add(new FirmwareUploaderPanel());
-		dialog.pack();
-		dialog.setLocationRelativeTo(mainFrame);
-
-		enableMenuBar(false);
-		dialog.addWindowListener(new WindowAdapter() {
-			@Override
-			public void windowClosing(WindowEvent e) {
-				enableMenuBar(true);
-			}
-		});
-
-		dialog.setVisible(true);
-	}
-
-	private JMenu createViewMenu() {
-		JMenu menu = new JMenu(Translator.get("MenuView"));
-		menu.setMnemonic('V');
-		
-		JMenuItem buttonZoomOut = new JMenuItem(Translator.get("MenuView.zoomOut"), KeyEvent.VK_MINUS);
-		buttonZoomOut.setAccelerator(KeyStroke.getKeyStroke(KeyEvent.VK_MINUS, SHORTCUT_CTRL));
-		buttonZoomOut.addActionListener((e) -> camera.zoom(1));
-		buttonZoomOut.setIcon(new UnicodeIcon("\uD83D\uDD0D-"));
-		menu.add(buttonZoomOut);
-
-		JMenuItem buttonZoomIn = new JMenuItem(Translator.get("MenuView.zoomIn"), KeyEvent.VK_EQUALS);
-		buttonZoomIn.setAccelerator(KeyStroke.getKeyStroke(KeyEvent.VK_EQUALS, SHORTCUT_CTRL));
-		buttonZoomIn.addActionListener((e) -> camera.zoom(-1));
-		buttonZoomIn.setIcon(new UnicodeIcon("\uD83D\uDD0D+"));
-		menu.add(buttonZoomIn);
-		
-		JMenuItem buttonZoomToFit = new JMenuItem(Translator.get("MenuView.zoomFit"), KeyEvent.VK_0);
-		buttonZoomToFit.setAccelerator(KeyStroke.getKeyStroke(KeyEvent.VK_0, SHORTCUT_CTRL));
-		buttonZoomToFit.addActionListener((e) -> camera.zoomToFit(myPaper.getPaperWidth(),myPaper.getPaperHeight()));
-		menu.add(buttonZoomToFit);
-
-		JCheckBoxMenuItem checkboxShowPenUpMoves = new JCheckBoxMenuItem(Translator.get("GFXPreferences.showPenUp"),GFXPreferences.getShowPenUp());
-		checkboxShowPenUpMoves.setAccelerator(KeyStroke.getKeyStroke(java.awt.event.KeyEvent.VK_M, SHORTCUT_CTRL));//"ctrl M" 
-		checkboxShowPenUpMoves.addActionListener((e) -> {
-			boolean b = GFXPreferences.getShowPenUp();
-			GFXPreferences.setShowPenUp(!b);
-		});
-		GFXPreferences.addListener((e)->{
-			checkboxShowPenUpMoves.setSelected ((boolean)e.getNewValue());
-		});
-		checkboxShowPenUpMoves.setIcon(new UnicodeIcon("\uD83D\uDC41"));
-		menu.add(checkboxShowPenUpMoves);
-
-		menu.add(createRenderStyleMenu());
-
-		return menu;
-	}
-
-	private JMenu createHelpMenu() {
-		JMenu menu = new JMenu(Translator.get("Help"));
-		menu.setMnemonic('H');//
-		
-		JMenuItem buttonViewLog = new JMenuItem(Translator.get("ShowLog"));
-		buttonViewLog.addActionListener((e) -> runLogPanel());		 
-		buttonViewLog.setAccelerator(KeyStroke.getKeyStroke(KeyEvent.VK_L, SHORTCUT_CTRL));//"ctrl L"
-		menu.add(buttonViewLog);
-		
-		JMenuItem buttonForums = createMenuItemBrowse(Translator.get("MenuForums"), "https://discord.gg/Q5TZFmB");
-		menu.add(buttonForums);
-
-		JMenuItem buttonDonation = createMenuItemBrowse(Translator.get("MenuItemPayPalDonation"), "https://www.marginallyclever.com/products/makelangelo-software/");
-		menu.add(buttonDonation);
-		
-		JMenuItem buttonCheckForUpdate = new JMenuItem(Translator.get("MenuUpdate"));
-		buttonCheckForUpdate.addActionListener((e) -> checkForUpdate(false));
-		menu.add(buttonCheckForUpdate);
-
-		menu.addSeparator();
-		
-		if (!isMacOS) {
-			JMenuItem buttonAbout = new JMenuItem(Translator.get("MenuAbout"));
-			buttonAbout.addActionListener((e) -> onDialogButton());
-			menu.add(buttonAbout);
-		}
-
-		
-		return menu;
-	}
-
-	public JMenuItem createMenuItemBrowse(String menuLabelAlreadyTranslated, String urlAsString) {
-		JMenuItem jmi = new JMenuItem(menuLabelAlreadyTranslated);
-		jmi.setToolTipText(urlAsString);
-		jmi.addActionListener((e) -> {
-			try {
-				if (Desktop.isDesktopSupported()) {
-					if (Desktop.getDesktop().isSupported(Desktop.Action.BROWSE)) {
-						java.awt.Desktop.getDesktop().browse(URI.create(urlAsString));
-					} else {
-						logger.error("Desktop.Action.BROWSE not supported. Can't open the browser to " + urlAsString);
-					}
-				} else {
-					logger.error("Desktop not supported. Can't open the browser to " + urlAsString);
-				}
-				
-			} catch (IOException ioe) {
-				logger.error("Can't open the browser to "+urlAsString, ioe);
-			}
-		});
-		return jmi;
-	}
-
-	private void onDialogButton() {
+	public void onDialogAbout() {
 		DialogAbout a = new DialogAbout();
 		a.display(mainFrame,MakelangeloVersion.VERSION, MakelangeloVersion.DETAILED_VERSION);
-	}
-
-	private void runLogPanel() {
-		LogPanel.runAsDialog(mainFrame);
 	}
 
 	/**
 	 * Parse <a href="https://github.com/MarginallyClever/Makelangelo/releases/latest">https://github.com/MarginallyClever/Makelangelo/releases/latest</a>
 	 * redirect notice to find the latest release tag.
 	 */
-	private void checkForUpdate(boolean announceIfFailure) {
+	public void checkForUpdate(boolean announceIfFailure) {
 		logger.debug("checking for updates...");
 		try {
 			URL github = new URL("https://github.com/MarginallyClever/Makelangelo-Software/releases/latest");
@@ -822,42 +325,7 @@ public final class Makelangelo {
 	 */
 	private void createRangeSlider(JPanel contentPane) {
 		logger.debug("  create range slider...");
-		JPanel bottomPanel = new JPanel(new BorderLayout());
-		rangeSlider = new RangeSlider();
-		rangeSlider.addChangeListener(this::onChangeSlider);
-		
-		labelRangeMax.setHorizontalAlignment(SwingConstants.RIGHT);
-		
-		labelRangeMin.setBorder(BorderFactory.createEmptyBorder(0, 5, 0, 5));
-		labelRangeMax.setBorder(BorderFactory.createEmptyBorder(0, 5, 0, 5));
-		
-		labelRangeMin.setFont(new Font(Font.MONOSPACED, Font.PLAIN, 12));
-		labelRangeMax.setFont(new Font(Font.MONOSPACED, Font.PLAIN, 12));
-		
-		Dimension d = labelRangeMin.getPreferredSize();
-		d.width=60;
-		labelRangeMin.setPreferredSize(d);
-		labelRangeMax.setPreferredSize(d);
-		
-		bottomPanel.add(labelRangeMin, BorderLayout.WEST);
-		bottomPanel.add(rangeSlider, BorderLayout.CENTER);
-		bottomPanel.add(labelRangeMax, BorderLayout.EAST);
-		
-		contentPane.add(bottomPanel, BorderLayout.SOUTH);
-	}
-	
-	/**
-	 * When the two-headed drawing start/end slider is moved
-	 * @param e {@link ChangeEvent} describing the move.
-	 */
-	private void onChangeSlider(ChangeEvent e) {
-        RangeSlider slider = (RangeSlider)e.getSource();
-        int bottom = slider.getValue();
-        int top = slider.getUpperValue();
-        myTurtleRenderer.setFirst(bottom);
-        myTurtleRenderer.setLast(top);
-        labelRangeMin.setText(Integer.toString(bottom));
-        labelRangeMax.setText(Integer.toString(top));
+		contentPane.add(rangeSlider, BorderLayout.SOUTH);
 	}
 
 	//  For thread safety this method should be invoked from the event-dispatching thread.
@@ -884,17 +352,14 @@ public final class Makelangelo {
 		} catch (IOException e) {
 			logger.warn("Can't load icon", e);
 		}
-		buildMenuBar();
-		
+
+		mainFrame.setJMenuBar(mainMenuBar);
 		mainFrame.setContentPane(createContentPane());
 		logger.debug("  make visible...");
 		mainFrame.setVisible(true);
 		mainFrame.setWindowSizeAndPosition();
 
 		camera.zoomToFit( Paper.DEFAULT_WIDTH, Paper.DEFAULT_HEIGHT);
-
-		openFileChooser = new OpenFileChooser(mainFrame);
-		openFileChooser.setOpenListener(this::openFile);
 
 		loadPaths();
 
@@ -910,7 +375,7 @@ public final class Makelangelo {
 				});
 			}
 			if (desktop.isSupported(Desktop.Action.APP_ABOUT)) {
-				desktop.setAboutHandler((e) -> onDialogButton());
+				desktop.setAboutHandler((e) -> onDialogAbout());
 			}
 		}
 	}
@@ -958,17 +423,7 @@ public final class Makelangelo {
 		SaveDialog.setLastPath( preferences.get(PREFERENCE_SAVE_PATH, FileAccess.getWorkingDirectory() ) );
 	}
 
-	private void saveFile() {
-		logger.debug("Saving vector file...");
-		try {
-			saveDialog.run(myTurtle, mainFrame);
-		} catch(Exception e) {
-			logger.error("Error while saving the vector file", e);
-			JOptionPane.showMessageDialog(mainFrame, Translator.get("SaveError") + e.getLocalizedMessage(), Translator.get("ErrorTitle"), JOptionPane.ERROR_MESSAGE);
-		}
-	}
-
-	private void setMainTitle(String title) {
+	public void setMainTitle(String title) {
 		String finalTitle = MakelangeloVersion.getFullOrLiteVersionStringRelativeToSysEnvDevValue();
 		if (!title.trim().isEmpty()) {
 			finalTitle = title + " - " + finalTitle;
@@ -980,10 +435,7 @@ public final class Makelangelo {
 		myTurtle = turtle;
 		myTurtleRenderer.setTurtle(turtle);
 		int top = turtle.history.size();
-		rangeSlider.setMinimum(0);
-		rangeSlider.setValue(0);
-		rangeSlider.setMaximum(top);
-		rangeSlider.setUpperValue(top);
+		rangeSlider.setLimits(0,top);
 	}
 
 	public Turtle getTurtle() {
@@ -1008,5 +460,43 @@ public final class Makelangelo {
 			Makelangelo makelangeloProgram = new Makelangelo();
 			makelangeloProgram.run();
 		});
+	}
+
+	public void saveGCode() {
+		logger.debug("Saving to gcode...");
+
+		SaveGCode save = new SaveGCode();
+		try {
+			int head = rangeSlider.getValue();
+			int tail = rangeSlider.getUpperValue();
+			save.run(getTurtle(), getPlotter(), mainFrame, head, tail);
+		} catch(Exception e) {
+			logger.error("Error while exporting the gcode", e);
+			JOptionPane.showMessageDialog(mainFrame, Translator.get("SaveError") + e.getLocalizedMessage(), Translator.get("ErrorTitle"), JOptionPane.ERROR_MESSAGE);
+		}
+	}
+
+	public Paper getPaper() {
+		return myPaper;
+	}
+
+	public Plotter getPlotter() {
+		return myPlotter;
+	}
+
+	public TurtleRenderer getTurtleRenderer() {
+		return myTurtleRenderer.getRenderer();
+	}
+
+	public Camera getCamera() {
+		return camera;
+	}
+
+	public void setTurtleRenderer(TurtleRenderer renderer) {
+		myTurtleRenderer.setRenderer(renderer);
+	}
+
+	public PlotterSettingsManager getPlotterSettingsManager() {
+		return plotterSettingsManager;
 	}
 }
